# [PERFORMANCE] Improve computation/memory on simple traversals