--- a/src/performance.spec.ts
+++ b/src/performance.spec.ts
@@ -22,7 +22,8 @@ describe('traversal performance tests', () => {
   describe.each([
     {n: 100,   expected: 5},
     {n: 1000,  expected: 50},
-    {n: 10000, expected: 5000},
+    {n: 10000, expected: 500},
+    {n: 100000, expected: 5000},
   ])('traversing $n times from single node cyclic graph', ({n, expected}) => {
     test(`should finish in ${expected}ms`, () => {
       G = bgraph.graph(...buildCyclicGraph(1));