Benchmarks (here Xn-Ye stands for a simple traversal `G.v().out().run()` where X is the number of nodes and Y is number of times out is called):

```
Before:
1n-100r: 1.0743703842163086ms
1n-1000r: 26.232311248779297ms
1n-10000r: 1058.5187377929688ms
10000n-1r: 20.17758846282959ms
100000n-1r: 107.87045478820801ms
1000000n-1r: 853.9200029373169ms

After:
1n-100r: 0.9706020355224609ms
1n-1000r: 5.8925933837890625ms
1n-10000r: 38.375250816345215ms
10000n-1r: 13.655800819396973ms
100000n-1r: 98.51250076293945ms
1000000n-1r: 626.1153287887573ms
```

## Changes:
- Use strict equalities in virtual machine instead of loose equalities to avoid performance loses from coercion.
- Use a linked list instead of an array when storing paths within Gremlin state to avoid memory cost of large shallowly copied arrays and computation cost of array iteration. Just the pointer references in these shallow copies was consuming a lot of memory and iterating through these arrays to perform a shallow copy was slowing down computation.